requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-postgres
psycopg[binary]
pgvector
pandas
prophet
openpyxl
reportlab
requests
python-dateutil
pymysql
redis
redis langchain-core